Move ajaxEvent register to config.js
