+++ ClassFileClient/assets/js/site.js
@@ -1,16 +1,19 @@
 import * as Route from "./common/routing.js";
 import * as Const from "./common/const.js";
-import { login } from "./auth/login.js";
-import { loadClass } from "./index.js";
+import { ajaxEvent } from "./common/config.js";
+import { login } from "./component/auth/login.js";
+import { loadClass } from "./component/index.js";
 
 function main() {
   Route.includeHTML();
   Route.addAuthHeader();
+  ajaxEvent();
 
   if (!Route.verifyAuth()) {
     return;
   }
 
+  // Đăng kí các function với URL html
   if (Route.checkPath(Const.Path.Login)) {
     login();
   } else if (Route.checkPath(Const.Path.Index)) {
@@ -18,47 +21,5 @@ function main() {
   }
 }
 
-// Đăng ký sự kiện Ajax
-$(document).ajaxSend(function (event, xhr, settings) {
-  if (
-    settings.type === Const.HttpMethod.POST ||
-    settings.type === Const.HttpMethod.PUT
-  )
-    Swal.fire({
-      title: Const.Message.Process,
-      allowEscapeKey: false,
-      allowOutsideClick: false,
-      timer: 2000,
-      didOpen: () => {
-        Swal.showLoading();
-      },
-    });
-});
-$(document).ajaxError(function (event, xhr, settings) {
-  let message = xhr.responseJSON?.message;
-  Swal.fire({
-    icon: "error",
-    title: Const.Message.Oops,
-    text: message,
-    footer: xhr.status + " - " + xhr.statusText,
-  });
-});
-$(document).ajaxComplete(function (event, xhr, settings) {
-  if (xhr.status === Const.HttpCode.UnAuthorized) {
-    Route.redirect(Const.Path.Login);
-  }
-  if (
-    settings.type === Const.HttpMethod.POST ||
-    settings.type === Const.HttpMethod.PUT
-  ) {
-    if (xhr.status === Const.HttpCode.Ok) {
-      Swal.fire({
-        icon: "success",
-        title: Const.Message.Success,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-    }
-  }
-});
+
 $(document).ready(main);
